requirements.txt
jittor==1.3.4.7
matplotlib==3.5.2
numpy==1.21.6
opencv-python==4.5.5.64
packaging==21.3
Pillow==9.1.1
scipy==1.7.3
tensorboardX==2.5
tqdm==4.64.0
typing_extensions==4.2.0
zipp==3.8.0
